src/test/java/com/antonylhz/shuati/random/ZigzagIterable2DListTest.java
package com.antonylhz.shuati.random;

import java.util.ArrayList;
import java.util.List;

import org.junit.Test;

import static org.junit.Assert.*;

public class ZigzagIterable2DListTest {
    @Test
    public void testSquare() {
        List<List<Integer>> data = new ArrayList<>();
        for (int i = 0; i < 5; i++) {
            List<Integer> list = new ArrayList<>();
            for (int j = 0; j < 5; j++) {
                list.add(i + 1);
            }
            data.add(list);
        }

        ZigzagIterable2DList.LListIterator iterator = (new ZigzagIterable2DList<Integer>(data)).iterator();
        assertTrue(iterator.hasNext());
        assertEquals(1, iterator.next());
        assertTrue(iterator.hasNext());
        assertEquals(2, iterator.next());
        assertTrue(iterator.hasNext());
        assertEquals(3, iterator.next());
        assertTrue(iterator.hasNext());
        assertEquals(4, iterator.next());
        assertTrue(iterator.hasNext());
        assertEquals(5, iterator.next());
        assertTrue(iterator.hasNext());
        assertEquals(1, iterator.next());
        assertTrue(iterator.hasNext());
        assertEquals(2, iterator.next());
        assertTrue(iterator.hasPrevious());
        assertEquals(1, iterator.previous());
    }
}